Module i3nator::types

The types in this module make up the structure of the project configuration files.

Example

The following is an examplary TOML configuration, which will be parsed into this modules types.

# i3nator project

# General configuration items
[general]
# Working directory to use
working_directory = "/path/to/my/working/directory"

# Name of the workspace the layout should be applied to
workspace = "1"

# Path to your layout-file
layout_path = "/path/to/my/layout.json"

# Alternatively, you can include the JSON-contents of the layout directly:
# layout = "{ ... }"

# List of applications to start
[[applications]]
command = "mycommand --with 'multiple args'"
working_directory = "/path/to/a/different/working/directory"

Structs

Application
The applications configuration.

ApplicationCommand
The command used for starting an application.

Config
This is the parent type defining the complete project configuration used by i3nator.

Exec
Commands to execute or keys to simulate after application startup.

General
The general configuration section.

Enums

ExecType
Defines how the commands in Exec should be interpreted.

Layout
This holds the layout, in multiple formats.