Crate i3nator

i3nator

i3nator is [Tmuxinator][gh-tmuxinator] for the [i3 window manager][i3wm].

It allows you to manage what are called "projects", which are used to easily restore saved i3 layouts (see [Layout saving in i3][i3wm-layout-saving]) and extending i3's base functionality by allowing you to automatically start applications too.

For detailed introductions, see the README.

License

DFW is licensed under either of

Apache License, Version 2.0, (http://www.apache.org/licenses/LICENSE-2.0)

MIT license (https://opensource.org/licenses/MIT)

at your option.

Modules

configfiles
Module consolidating common functionality between projects and layouts.

errors
Errors, using error-chain.

layouts
Module for layout handling.

projects
Module for project handling.

types
The types in this module make up the structure of the project configuration files.