Trait i3nator::configfiles::ConfigFile

pub trait ConfigFile: Sized {
    fn copy<S: AsRef<OsStr> + ?Sized>(&self, new_name: &S) -> Result<Self>;

    fn create<S: AsRef<OsStr> + ?Sized>(name: &S) -> Result<Self>;

    fn create_from_template<S: AsRef<OsStr> + ?Sized>(
        name: &S, 
        template: &[u8]
    ) -> Result<Self>;

    fn delete(&self) -> Result<()>;

    fn from_path<P: AsRef<Path> + ?Sized>(path: &P) -> Result<Self>;

    fn list() -> Vec<OsString>;

    fn name(&self) -> String;

    fn open<S: AsRef<OsStr> + ?Sized>(name: &S) -> Result<Self>;

    fn path(&self) -> PathBuf;

    fn prefix() -> &'static OsStr;

    fn rename<S: AsRef<OsStr> + ?Sized>(&self, new_name: &S) -> Result<Self>;

    fn verify(&self) -> Result<()>;
}

Helping type to consolidate common functionality between projects and layouts.

Required methods

fn copy<S: AsRef<OsStr> + ?Sized>(&self, new_name: &S) -> Result<Self>

Create a copy of the current configfile, that is a copy of the configuration file on disk, with a name of new_name.

This will keep the same prefix.

Parameters

new_name: A OsStr that is the name of the destination configfile.

Returns

A Result which is:

Ok: an instance of ConfigFile for the new configfile.

Err: an error, e.g. if a configfile with new_name already exists or copying the file failed.

fn create<S: AsRef<OsStr> + ?Sized>(name: &S) -> Result<Self>

Create a configfile given a name.

This will not create the configuration file, but it will ensure a legal XDG path with all directories leading up to the file existing.

If you want to pre-fill the configuration file with a template, see ConfigFile::create_from_template.

Parameters

name: A OsStr naming the configuration file on disk.

Returns

A Result which is:

Ok: an instance of ConfigFile for the given name.

Err: an error, e.g. if the configfile already exists or couldn’t be created.

fn create_from_template<S: AsRef<OsStr> + ?Sized>(
name: &S,
template: &[u8]
) -> Result<Self>

Create a configfile given a name, pre-filling it with a given template.

See ConfigFile::create for additional information.

Parameters

name: A OsStr naming the the configuration file on disk.

template: A byte-slice which will be written to the configuration file on disk.

Returns

A Result which is:

Ok: an instance of ConfigFile for the given name with the contents of template.

Err: an error, e.g. if the configfile already exists or couldn’t be created.

fn delete(&self) -> Result<()>

Delete this configfile from disk.

Returns

A Result which is:

Ok: nothing (()).

Err: an error if deleting the file failed.

fn from_path<P: AsRef<Path> + ?Sized>(path: &P) -> Result<Self>

Opens an existing configfile for a given path.

This will not impose any XDG conventions, but rather allows to open a configuration from any path.

See ConfigFile::open if you want to open a configfile by name and prefix.

Parameters

path: A Path specifiying the configuration file on disk.

Returns

A Result which is:

Ok: an instance of ConfigFile for the given path.

Err: an error, e.g. if the file does not exist.

fn list() -> Vec<OsString>

Get a list of all configfile names.

This will check the current users XDG base directories for configuration files, and return a list of their names for use with e.g. ConfigFile::open.

fn name(&self) -> String

Returns the name of this configfile.

As represented by the stem of the filename on disk.

fn open<S: AsRef<OsStr> + ?Sized>(name: &S) -> Result<Self>

Opens an existing configfile using a name.

This will search for a matching configfile in the XDG directories.

See ConfigFile::from_path if you want to open a configfile using any path.

Parameters

name: A OsStr naming the configuration file on disk.

Returns

A Result which is:

Ok: an instance of ConfigFile for the given name.

Err: an error, e.g. if the file does not exist.

fn path(&self) -> PathBuf

Returns the path to the configfile.

fn prefix() -> &'static OsStr

Return the prefix associated with this type of configfile.

fn rename<S: AsRef<OsStr> + ?Sized>(&self, new_name: &S) -> Result<Self>

Rename the current configfile.

Parameters

new_name: A OsStr that is the name of the destination configfile.

Returns

A Result which is:

Ok: an instance of ConfigFile for the renamed configfile.

Err: an error, e.g. if a configfile with new_name already exists or renaming the file failed.

fn verify(&self) -> Result<()>

This verifies the project’s configuration, without storing it in the current project instance.

Returns

A Result which is:

Ok: nothing (()) if the verification succeeded.

Err: an error if the configuration could not be parsed with details on what failed.