Enum i3nator::errors::ErrorKind

pub enum ErrorKind {
    I3EstablishError(EstablishError),
    I3MessageError(MessageError),
    IoError(Error),
    Utf8Error(Utf8Error),
    TomlError(Error),
    Msg(String),
    CantBeImplemented(String),
    CommandSplittingFailed(String),
    ConfigExists(String, String),
    EditorNotFound,
    InvalidUtF8Path(String),
    PathDoesntExist(String),
    TextOrKeyInputFailed,
    UnknownConfig(String, String),
    // some variants omitted
}

The kind of an error.

Variants

I3EstablishError(EstablishError)

Error caused by i3ipc, on establishing a connection.

I3MessageError(MessageError)

Error caused by i3ipc, on sending a message.

IoError(Error)

Error mapping to std::io::Error.

Utf8Error(Utf8Error)

Error mapping to std::str::Utf8Error.

TomlError(Error)

Error caused by toml, on deserializing using Serde.

Msg(String)

A convenient variant for String.

CantBeImplemented(String)

An error that occurs if a trait-function is called that cannot be implemented.

(This is pretty unclean but is currently required as ConfigFileImpl cannot implement any of the static functions of the ConfigFile trait since it is missing the required prefix.)

CommandSplittingFailed(String)

An error that can occur when splitting a string into a ApplicationCommand.

ConfigExists(String, String)

An error that occurs if a project under the same name already exists.

EditorNotFound

An error that occurs when the default editor is not specified.

One of the environment variables $VISUAL or $EDITOR has to be set.

InvalidUtF8Path(String)

An error that occurs when a Path (i.e. OsStr) cannot be converted to UTF8.

PathDoesntExist(String)

An error that occurs if a specified path does not exist.

TextOrKeyInputFailed

An error that occurs if text or key-presses could not be input into an application.

UnknownConfig(String, String)

An error that occurs if a project does not exist under a specified name.

Implementations

impl ErrorKind

pub fn description(&self) -> &str

A string describing the error kind.